platform/requirements.txt
python-roborock>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# 音频感官层 (AudioSensor)
sounddevice>=0.5.0
webrtcvad>=2.0.10
numpy>=1.26.0

# 音频效应器 (AudioEffector)
edge-tts>=7.0.0
soundfile>=0.12.1

# 摄像头采集 (Camera)
opencv-python-headless>=4.8.0

# 激光雷达 (Lidar / LD06，CP2102 USB-TTL 串口)
pyserial>=3.5

# SLAM 建图 (breezyslam，Hector SLAM 风格扫描匹配，无需里程计)
breezyslam>=0.8.0

# 电源监测 (INA219 电流/电压传感器，I2C)
pi-ina219>=1.3.0
